Create file of given size if !exist
Allow the user to create a file backstore by specifying a new file name
and a size. We create the file and blow it up to the size (will be fast if
sparse files are in use.

Handle rename of convert_human_to_bytes -> human_to_bytes

Signed-off-by: Andy Grover <agrover@redhat.com>

Merged by Jerome Martin <tramjoe.merin@gmail.com>
Conflicts:
	targetcli/ui_backstore.py