+ command options for sessions, close #11
Made listing the session details optional,
option to show only a specific session,
documented the session command,
check given options.

The bare list doesn't show much specific for the session,
only the alias, sid, type and state.
Displaying the inititatorname would be possible, but it is too
long to fit the entry on one line then.

Signed-off-by: Johannes Dewender <github@JonnyJD.net>