Update version for proper pkg versioning
Unfortunately I made a mistake, and did not handle version with 'rc' in it
as specified here:

http://fedoraproject.org/wiki/Packaging:NamingGuidelines#Pre-Release_packages

I'm incrementing the version to 2.1.<fb-ver> to get version sequencing
right. It's a little uncool for me to be mixing my developer and packager
roles here, but I think I can rationalize a version bump, given the
tag and WWN improvements recently.

Signed-off-by: Andy Grover <agrover@redhat.com>