Enhance summary line for TPGs and ACLs based on auth options
LIO has many different ways to configure authentication. This might be
the most confusing part of using LIO. The key bit is that LIO *always*
wants an ACL for an initiator -- the question is if these are manually
created, or created on-the-fly. This is controlled by
"attrib/generate_node_acls".

Change TPG summary line to always show the state of generate_node_acls.
Furthermore, make the status green if the settings for a particular
auth configuration are set:

* authentication=0
  No authentication is performed, so anyone can connect. Since iqns are
  trivially forgeable, manual ACLs are silly. The user probably also wants
  generate_node_acls=1.

* authentication=1, generate_node_acls=1
  ACLs are generated and auth settings are at the TPG level. We go green
  if CHAP userid and password are also set for the TPG. ACL summary tries
  to direct the user to look at the TPG. Also indicate if mutual auth is
  configured.

* authentication=1, generate_node_acls=0
  Auth at the ACL level. Manually created ACLs. Try to tell the user to
  look at the ACL-level, and go green only if userid and password are set.
  Also indicate if mutual auth is configured.

Signed-off-by: Andy Grover <agrover@redhat.com>