Allow creation of MappedLUN or LUN to also create storage object
If argument is not a valid storageobject path (or lun, for nodeacl),
also check to see if it's a file or block device, and attempt to
create a new storage object backed by it, and proceed from there.

Enable path autocompletion to go along with this.

This allows mapping a block device to an existing NodeACL in one command.

Signed-off-by: Andy Grover <agrover@redhat.com>