New "configure" mode in targetcli
This is an early preview of the next-generation targetcli based on the new
rtslib.Config API and the lio shell from the rtsli lio-config branch.

This mode allows editing a candidate configuration without impacting the
running system. This candidate configuration can then either be commited or
discarded at will. If commited, it will be applied to the running system and
saved as the new startup configuration.

Other features include loading a configuration from file, undo support,
rollback support, configuration backups and more.

This new mode is using a subshell not based anymore on configshell.
The plan is to introduce this new subshell in the next release, deemed as a
transition release, and eventually replace entirely the current targetcli UI,
when it will have reached parity with the current configshell both in terms of
features and quality of user interaction (colors, syntetic view of the system,
auto-completion, etc.). A basic standalone version of this shell is included
as targetcli-ng.

More information can be found in the target-devel mailing-list, notably the
"[ANN] LIO userspace improvements" thread dated 2014-02-27:

http://thread.gmane.org/gmane.linux.scsi.target.devel/5838/focus=5847

The code was manually ported over from the rtslib lio-config branch.