Fix initscript to accept missing fabric modules
It is possible that we include spec files with rtlib for fabric modules not
currently installed. A current example is the usb_gadget fabric module.

The target initscript will simply issue a warning when a fabric module
referenced in a spec file cannot be loaded.

Also, on stop, the initscript now tries to do as much as possible, even
if unloading something fails along the way.