Document target parameters and attributes
Parameters read from the config file system have the same generic
description that does not provide much information:

  > get parameter
  MaxBurstLength=262144
  ---------------------
  The MaxBurstLength parameter.

This patch provides a description for parameters and attributes by
adding 'ui_desc_parameters' and 'ui_desc_attributes' class members to
the UITPG, UINodeACL and UIStorageObject classes.

The description is read when the parameter is registered with
define_config_group_param(). It is displayed when the user invokes
`get parameter` or `get attribute`:

  > get parameter
  MaxBurstLength=262144
  ---------------------
  Maximum SCSI data payload in bytes in a Data-In or a solicited Data-Out iSCSI sequence.

This patch also declares the types of parameters. Thus, targetcli can
more efficiently check the parameters when they are set.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>