Add enabling service to man page quickstart section
One common issue is that we require the distro to restore settings on
reboot, and how to enable this can vary based on what init system is in
use. One additional gotcha is that we pushed this down into rtslib, so
a user looking for init scripts or service files in a distro targetcli
package won't find them.

Try to document this better in targetcli.8 so admins will have an easier
time.

Signed-off-by: Andy Grover <agrover@redhat.com>