Revert checking for if TCMUService1 exists
One issue I was seeing in development was long timeouts if the tcmu-runner
dbus service file was installed but tcmu-runner wasn't started. This was
an attempt to fix this.

However, having the .service file and not having tcmu-runner there to
be started should never happen, so we don't need to do this.

Signed-off-by: Andy Grover <agrover@redhat.com>