Allow WWN to be specified when creating storage objects
rtslib's API supports passing in the WWN for a new storage object, but
up until now targetcli's "create" commands have not had this as an option
Add this.

This will make it easier to configure things like multipath,
which uses wwn to know if different block devices actually refer to the
same underlying storage.

Reported-by: Paul Cuzner <pcuzner@redhat.com>
Signed-off-by: Andy Grover <agrover@redhat.com>