Clear backstore cache on refresh
rtslib creates a backstore cache from configfs as a global variable.
The targetcli refresh function needs to clear out said cache, or you
end up with errors such as the following from an active shell if
you create a new backstore from a different shell.

/> refresh
Storage object fileio/test not found

In order to work this depends on pull request #87 from rtslib_fb,
however it is written so that it won't break if an older version
of the library is still in use.

Signed-off-by: Amartey Pearson <apearson@us.ibm.com>