Allow tcmu to be enabled after creation
For TCMU we are going to be adding more control settings,
so instead of having to modify targetcli and rtslib for
each one, this allows us to enable the device after
creation. We can then set whatever settings we want
after creation without having to update the 2 packages.

It will also allow the enable step to fail, but not
destroy the entire object in the kernel so the user
can still see it, the lun exports are still setup,
and the user can just retry the enable call later.

Signed-off-by: Mike Christie <mchristi@redhat.com>