Use temp file objects for temporary storage area
happen to see yet another issue after we switch to stringIO/bytesIO:

$ targetcli ls
'unicode' does not have the buffer interface

After which I felt like dealing with StringIO/BytesIO is like a
wild goose chase and we are after all attempting to deal with python
incompatible apis.

Technically speaking, the rtslib and configshell libraries are expecting a
string IO stream, but then for python2 with ByteIO() we are passing a
bytestream, and stringIO() is behaving a bit different in python2 than
expected as explained in my previous patch.

Lets simply switch to temporary file in the most secure manner possible,
opening files in string mode will assure the compatibility across the platforms.

At the end we wish to have a consistent behaviour across all python versions.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>