Improve the handling of stdout and stderr in daemon mode
In daemon mode, the STDOUT and STDERR messages were aggregated into a
single text file which was then relayed back to the caller. There was no
differentiation between the STDOUT and STDERR which caused it to return
0 and not log any STDERR even in command failure case. Following is an
example:

$ targetcli get global auto_use_daemon
auto_use_daemon=true

$ targetcli err; echo $?
Command not found err
0

The fix is to create separate temp files for stdout and stderr.

After the fix it returns error correctly (1):

$ targetcli get global auto_use_daemon
auto_use_daemon=true

$ targetcli err; echo $?
Command not found err
1

Signed-off-by: Joshua Brown <joshua.f.brown@oracle.com>
Signed-off-by: Gulam Mohamed <gulam.mohamed@oracle.com>